Rename endpoint-allocator to endpoint-id-allocator.
