+++ examples/all-devices-app/all-devices-common/devices/endpoint-id-allocator/BUILD.gn
@@ -14,20 +14,18 @@
 
 import("//build_overrides/chip.gni")
 
-import("${chip_root}/build/chip/chip_test_suite.gni")
-import("${chip_root}/build/chip/tests.gni")
+source_set("endpoint-id-allocator") {
+  sources = [
+    "ConsecutiveEndpointIdAllocator.h",
+    "DynamicEndpointIdAllocator.cpp",
+    "DynamicEndpointIdAllocator.h",
+    "EndpointIdAllocator.h",
+  ]
 
-if (chip_build_tests) {
-  chip_test_suite("tests") {
-    output_name = "TestEndpointIdAllocator"
+  public_deps = [ "${chip_root}/src/lib" ]
 
-    test_sources = [ "TestEndpointIdAllocator.cpp" ]
-
-    cflags = [ "-Wconversion" ]
-
-    public_deps = [
-      "${chip_root}/examples/all-devices-app/all-devices-common/devices/endpoint-allocator",
-      "${chip_root}/src/lib/support:testing",
-    ]
-  }
+  public_configs = [
+    "${chip_root}/examples/all-devices-app/all-devices-common/devices:includes",
+    "${chip_root}/src:includes",
+  ]
 }
