+++ examples/all-devices-app/all-devices-common/devices/endpoint-id-allocator/ConsecutiveEndpointIdAllocator.h
@@ -16,7 +16,7 @@
 
 #pragma once
 
-#include <devices/endpoint-allocator/EndpointIdAllocator.h>
+#include <devices/endpoint-id-allocator/EndpointIdAllocator.h>
 
 namespace chip::app {
