+++ examples/all-devices-app/all-devices-common/devices/endpoint-id-allocator/DynamicEndpointIdAllocator.cpp
