+++ examples/all-devices-app/all-devices-common/devices/endpoint-id-allocator/tests/BUILD.gn
@@ -14,18 +14,20 @@
 
 import("//build_overrides/chip.gni")
 
-source_set("endpoint-allocator") {
-  sources = [
-    "ConsecutiveEndpointIdAllocator.h",
-    "DynamicEndpointIdAllocator.cpp",
-    "DynamicEndpointIdAllocator.h",
-    "EndpointIdAllocator.h",
-  ]
+import("${chip_root}/build/chip/chip_test_suite.gni")
+import("${chip_root}/build/chip/tests.gni")
 
-  public_deps = [ "${chip_root}/src/lib" ]
+if (chip_build_tests) {
+  chip_test_suite("tests") {
+    output_name = "TestEndpointIdAllocator"
 
-  public_configs = [
-    "${chip_root}/examples/all-devices-app/all-devices-common/devices:includes",
-    "${chip_root}/src:includes",
-  ]
+    test_sources = [ "TestEndpointIdAllocator.cpp" ]
+
+    cflags = [ "-Wconversion" ]
+
+    public_deps = [
+      "${chip_root}/examples/all-devices-app/all-devices-common/devices/endpoint-id-allocator",
+      "${chip_root}/src/lib/support:testing",
+    ]
+  }
 }
