+++ examples/all-devices-app/all-devices-common/devices/interface/BUILD.gn
@@ -22,7 +22,7 @@ source_set("device-interface") {
   ]
 
   public_deps = [
-    "${chip_root}/examples/all-devices-app/all-devices-common/devices/endpoint-allocator",
+    "${chip_root}/examples/all-devices-app/all-devices-common/devices/endpoint-id-allocator",
     "${chip_root}/src/app/clusters/descriptor",
     "${chip_root}/src/data-model-providers/codedriven",
     "${chip_root}/src/lib/core:error",
